create gcp_keys in correct dir
currently the instructions say to run anthos/create_service_accounts.sh .  however, the script as written will then create gcp_keys subdir in the current dir "." which will be the top-level dir.  in contrast, the manual instructions say to create the gcp_keys subdir under anthos/.  this is inconsistent.

having the script create the gcp_keys subdir in a location that depends on where the script is called from, is error-prone.  rather, it should be bullet-proofed to always create gcp_keys in the "right" location.

with my change, gcp_keys will always be created in the parent dir of the script itself.